requirements.txt
# Core dependencies
anthropic>=0.39.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# CLI & UI
rich>=13.0.0
click>=8.1.0

# Web scraping & HTTP
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0
lxml>=5.0.0

# Excel generation
openpyxl>=3.1.0

# Utilities
tenacity>=8.2.0
python-dateutil>=2.8.0

# Development (optional)
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
